improved extensibility by creating example of how multiple custom platforms can be created
